Split editor.js into multiple files; more systematic naming conventions for files
