+++ app/src/index.html
@@ -9,14 +9,23 @@
 
 <body>
     <div class="flex flex-col h-screen">
-        <%= require("html-loader!./navbar/navbar.html").default %>
+        <%= require("html-loader!./navbar/UiNavbar.html").default %>
         <div class="grow h-0" id="screen-swapper"></div>
     </div>
 
-    <%= require("html-loader!./screens/login/ui/login.html").default %>
-    <%= require("html-loader!./screens/create-account/ui/create-account.html").default %>
-    <%= require("html-loader!./screens/dashboard/ui/dashboard.html").default %>
-    <%= require("html-loader!./screens/editor/ui/editor.html").default %>
+    <!-- TODO: Figure out to make webpack render templates recursively -->
+
+    <%= require("html-loader!./screens/login/ui/UiLoginScreen.html").default %>
+
+    <%= require("html-loader!./screens/create-account/ui/UiCreateAccountScreen.html").default %>
+
+    <%= require("html-loader!./screens/dashboard/ui/UiDashboardScreen.html").default %>
+    <%= require("html-loader!./screens/dashboard/ui/UiPresentationItem.html").default %>
+
+    <%= require("html-loader!./screens/editor/ui/UiEditorScreen.html").default %>
+    <%= require("html-loader!./screens/editor/ui/comment-sidebar/UiComment.html").default %>
+    <%= require("html-loader!./screens/editor/ui/thumbnail-sidebar/UiThumbnail.html").default %>
+    <%= require("html-loader!./screens/editor/ui/timeline/UiTimelineVersion.html").default %>
 </body>
 
 </html>
