+++ app/src/screens/dashboard/ui/UiPresentationItem.html
@@ -1,0 +1,46 @@
+<template id="presentation-template">
+    <div class="m-2 max-h-80 min-h-min content-center rounded bg-zinc-200 p-2 hover:cursor-pointer">
+        <!-- title text -->
+        <div class="my-auto flex min-h-[3rem] justify-center text-center">
+            <h2 class="presentation-title my-auto h-min font-bold">
+                Project title
+            </h2>
+        </div>
+        <!-- thumbnail -->
+        <div class="id-presentation-thumbnail relative my-2 flex h-20 w-full justify-center">
+            <div class="id-thumbnail-placeholder relative h-full w-min">
+                <!-- Heroicons photograph -->
+                <svg xmlns="http://www.w3.org/2000/svg" class="h-full w-auto text-gray-800"
+                    fill="none"
+                    viewBox="0 0 24 24"
+                    stroke="currentColor"
+                >
+                    <path
+                        stroke-linecap="round"
+                        stroke-linejoin="round"
+                        stroke-width="2"
+                        d="M4 16l4.586-4.586a2 2 0 012.828 0L16 16m-2-2l1.586-1.586a2 2 0 012.828 0L20 14m-6-6h.01M6 20h12a2 2 0 002-2V6a2 2 0 00-2-2H6a2 2 0 00-2 2v12a2 2 0 002 2z"
+                    />
+                </svg>
+                <!-- Heroicons dots -->
+                <svg
+                    xmlns="http://www.w3.org/2000/svg"
+                    class="absolute bottom-0 right-0 h-6 w-7 rounded bg-slate-500 px-1 text-white hover:bg-slate-700"
+                    viewBox="0 0 20 20"
+                    fill="currentColor"
+                >
+                    <path
+                        d="M6 10a2 2 0 11-4 0 2 2 0 014 0zM12 10a2 2 0 11-4 0 2 2 0 014 0zM16 12a2 2 0 100-4 2 2 0 000 4z"
+                    />
+                </svg>
+            </div>
+            <!-- /thumbnail -->
+        </div>
+        <!-- description text -->
+        <div class="text-center">
+            <h2 class="presentation-description italic">
+                Project description
+            </h2>
+        </div>
+    </div>
+</template>
