+++ app/src/screens/dashboard/ui/UiPresentationItem.js
@@ -1,0 +1,25 @@
+import { cloneDomTemplate } from "../../../common/ui/dom-utils.js";
+import { uiScreenSwapper } from "../../uiScreenSwapper.js";
+
+class UiPresentationItem {
+    constructor(screen, presentation) {
+        this.screen = screen;
+
+        this.el = cloneDomTemplate("#presentation-template");
+        this.el.addEventListener("click", () => this.onClick());
+
+        this.presentation = presentation;
+
+        this.titleEl = this.el.querySelector(".presentation-title");
+        this.titleEl.textContent = presentation.title;
+
+        this.descriptionEl = this.el.querySelector(".presentation-description");
+        this.descriptionEl.textContent = presentation.description;
+    }
+
+    onClick() {
+        uiScreenSwapper.loadScreen("editor", {presentation: this.presentation.appwriteId});
+    }
+}
+
+export { UiPresentationItem };
