--- app/src/screens/dashboard/ui/dashboard.html
@@ -1,79 +1,0 @@
-<template id="dashboard-screen-template">
-    <div>
-        <div class="relative grow overflow-y-scroll bg-zinc-50">
-            <h1 class="my-5 mx-5 font-bold">Authored Presentations</h1>
-            <div class="id-presentation-list my-5 mx-5 grid grid-cols-4 flex-col">
-                <!-- #presentation-template gets instantiated here. -->
-            </div>
-        </div>
-    
-        <!-- presentation creation -->
-        <div class="flex h-16 items-center justify-center gap-4 bg-zinc-200">
-            <input class="id-presentation-title-input rounded p-2 focus:outline-none" type="text" placeholder="Title" />
-            <input class="id-presentation-description-input rounded p-2 focus:outline-none" type="text" placeholder="Description" />
-            <button class="id-add-presentation-button rounded-full bg-slate-500 py-2 px-4 font-bold text-white hover:bg-slate-700">
-                Add Project
-            </button>
-            <!-- /presentation creation -->
-        </div>
-    </div>
-</template>
-
-<template id="presentation-template">
-    <div
-        class="m-2 max-h-80 min-h-min content-center rounded bg-zinc-200 p-2 hover:cursor-pointer"
-    >
-        <!-- title text -->
-        <div
-            class="my-auto flex min-h-[3rem] justify-center text-center"
-        >
-            <h2 class="presentation-title my-auto h-min font-bold">
-                Project title
-            </h2>
-        </div>
-        <!-- thumbnail -->
-        <div
-            class="relative my-2 flex h-20 w-full justify-center"
-            id="presentation-thumbnail"
-        >
-            <div
-                class="relative h-full w-min"
-                id="thumbnail-placeholder"
-            >
-                <!-- Heroicons photograph -->
-                <svg
-                    xmlns="http://www.w3.org/2000/svg"
-                    class="h-full w-auto text-gray-800"
-                    fill="none"
-                    viewBox="0 0 24 24"
-                    stroke="currentColor"
-                >
-                    <path
-                        stroke-linecap="round"
-                        stroke-linejoin="round"
-                        stroke-width="2"
-                        d="M4 16l4.586-4.586a2 2 0 012.828 0L16 16m-2-2l1.586-1.586a2 2 0 012.828 0L20 14m-6-6h.01M6 20h12a2 2 0 002-2V6a2 2 0 00-2-2H6a2 2 0 00-2 2v12a2 2 0 002 2z"
-                    />
-                </svg>
-                <!-- Heroicons dots -->
-                <svg
-                    xmlns="http://www.w3.org/2000/svg"
-                    class="absolute bottom-0 right-0 h-6 w-7 rounded bg-slate-500 px-1 text-white hover:bg-slate-700"
-                    viewBox="0 0 20 20"
-                    fill="currentColor"
-                >
-                    <path
-                        d="M6 10a2 2 0 11-4 0 2 2 0 014 0zM12 10a2 2 0 11-4 0 2 2 0 014 0zM16 12a2 2 0 100-4 2 2 0 000 4z"
-                    />
-                </svg>
-            </div>
-            <!-- /thumbnail -->
-        </div>
-        <!-- description text -->
-        <div class="text-center">
-            <h2 class="presentation-description italic">
-                Project description
-            </h2>
-        </div>
-    </div>
-</template>
