+++ app/src/screens/editor/ui/UiEditorScreen.js
@@ -1,0 +1,40 @@
+import { initData, terminateData } from "../model/data.js";
+import { UiRestrictedScreen } from "../../UiRestrictedScreen.js";
+import { UiThumbnailBar } from "./thumbnail-sidebar/UiThumbnailBar.js";
+import { UiContentCenter } from "./UiContentCenter.js";
+import { UiTimeline } from "./timeline/UiTimeline.js";
+import { UiRightSidebar } from "./comment-sidebar/UiRightSidebar.js";
+import { uiScreenRegistry } from "../../uiScreenRegistry.js";
+
+// TODO: We probably won't want to inherit from restricted screen, since people
+// should be able to add comments without an account? That doesn't work at the
+// moment, so for now this is probably okay.
+class UiEditorScreen extends UiRestrictedScreen {
+    static NAME = "editor";
+
+    constructor(screenParameters) {
+        super("#editor-screen-template", screenParameters);
+    }
+
+    initRestricted() {
+        initData(this.screenParameters.presentation);
+
+        this.thumbnailBar = new UiThumbnailBar(this);
+        this.contentCenter = new UiContentCenter(this);
+        this.timeline = new UiTimeline(this);
+        this.rightSideBar = new UiRightSidebar(this);
+    }
+
+    terminateRestricted() {
+        this.rightSideBar.terminate();
+        this.timeline.terminate();
+        this.contentCenter.terminate();
+        this.thumbnailBar.terminate();
+
+        terminateData();
+    }
+}
+
+uiScreenRegistry.add(UiEditorScreen);
+
+export { UiEditorScreen };
