+++ app/src/screens/editor/ui/comment-sidebar/UiComment.html
@@ -1,0 +1,17 @@
+<template id="comment-template">
+    <div class="h-auto w-full bg-slate-400 p-2">
+        <span class="font-bold text-orange-600 comment-author"></span>
+        <span class="comment-text"></span>
+    </div>
+</template>
+
+<!-- TODO: Thread-Template fertigstellen -->
+<template id="thread-template">
+    <!-- thread container -->
+    <div class="h-60 w-full flex-none items-center justify-start bg-black opacity-50">
+        <!-- TODO: erstellen eines Thread-Haupt-Kommentar-Templates -->
+        <div></div>
+        <!-- TODO: erstellen eines Thread-Antwort-Kommentar-Templates -->
+        <div></div>
+    </div>
+</template>
