+++ app/src/screens/editor/ui/comment-sidebar/UiRightSidebar.js
@@ -1,0 +1,17 @@
+import { UiCommentInputFields } from "./UiCommentInputFields.js";
+import { UiCommentList } from "./UiCommentList.js";
+
+class UiRightSidebar {
+    constructor(screen) {
+        this.el = screen.el.querySelector(".id-sidebar-right");
+
+        this.commentList = new UiCommentList(screen);
+        this.commentInputFields = new UiCommentInputFields(screen);
+    }
+
+    terminate() {
+        this.commentList.terminate();
+    }
+}
+
+export { UiRightSidebar };
