+++ app/src/screens/editor/ui/thumbnail-sidebar/UiThumbnail.html
@@ -1,0 +1,26 @@
+<template id="thumbnail-template">
+    <button class="transition p-4 flex flex-col space-y-2 rounded-xl bg-slate-800 hover:bg-slate-400 active:bg-slate-900 selected:bg-amber-500 text-white selected:text-black">
+        <div class="flex space-x-2 items-center">
+            <div class="text-sm opacity-80 page-number">12</div>
+            <div class="flex justify-center grow"><canvas></canvas></div>
+        </div>
+        <div class="flex space-x-3 items-center justify-center">
+            <!-- heroicons: chat -->
+            <span class="flex items-center space-x-1" title="Total number of comments">
+                <svg xmlns="http://www.w3.org/2000/svg" class="h-6 w-6" fill="none" viewBox="0 0 24 24" stroke="currentColor"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M8 12h.01M12 12h.01M16 12h.01M21 12c0 4.418-4.03 8-9 8a9.863 9.863 0 01-4.255-.949L3 20l1.395-3.72C3.512 15.042 3 13.574 3 12c0-4.418 4.03-8 9-8s9 3.582 9 8z" /></svg>
+                <span class="num-comments">7</span>
+            </span>
+            <!-- heroicons: exclamation-circle -->
+            <span class="flex items-center space-x-1" title="Number of unresolved threads">
+                <svg xmlns="http://www.w3.org/2000/svg" class="h-6 w-6" fill="none" viewBox="0 0 24 24" stroke="currentColor"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M12 8v4m0 4h.01M21 12a9 9 0 11-18 0 9 9 0 0118 0z" /></svg>
+                <span class="num-unresolved-threads">2</span>
+            </span>
+            <!-- TODO: Do we want to keep this? Tracking unread items seems a bit painful to implement ... -->
+            <!-- heroicons: sparkles -->
+            <span class="flex items-center space-x-1" title="Number of unread comments">
+                <svg xmlns="http://www.w3.org/2000/svg" class="h-6 w-6" fill="none" viewBox="0 0 24 24" stroke="currentColor"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M5 3v4M3 5h4M6 17v4m-2-2h4m5-16l2.286 6.857L21 12l-5.714 2.143L13 21l-2.286-6.857L5 12l5.714-2.143L13 3z" /></svg>
+                <span class="num-unread-comments">1</span>
+            </span>
+        </div>
+    </button>
+</template>
