+++ app/src/screens/editor/ui/timeline/UiTimelineVersion.html
@@ -1,0 +1,5 @@
+<template id="version-template">
+    <button class="transition p-4 rounded-xl bg-transparent hover:bg-slate-400 active:bg-slate-900 selected:bg-amber-500 text-white selected:text-black">
+        <span class="label">12</span>
+    </button>
+</template>
