+++ app/src/screens/import-screens.js
@@ -1,0 +1,5 @@
+// Import to add the screens to the uiScreenRegistry.
+import "./login/ui/UiLoginScreen.js";
+import "./create-account/ui/UiCreateAccountScreen.js";
+import "./dashboard/ui/UiDashboardScreen.js";
+import "./editor/ui/UiEditorScreen.js";
